+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - added o4-mini model in Agent
 - properly handling user input when tool call is denied by user
 - added model selector for Agent model to Project bar
+- using env vars for Agent LLM providers (API keys, etc.)
 
 ## [0.9.0]